+++ util.h
@@ -5,16 +5,16 @@
 
 //------------------------------------------------------------------------------
 
-void   mid2(double *, const double *, const double *);
-void   mid4(double *, const double *, const double *,
-                      const double *, const double *);
+void  mid2(float *, const float *, const float *);
+void  mid4(float *, const float *, const float *,
+                    const float *, const float *);
 
-double lerp1(double, double, double);
-double lerp2(double, double, double, double, double, double);
+float lerp1(float, float, float);
+float lerp2(float, float, float, float, float, float);
 
-void   slerp1(double *, const double *, const double *, double);
-void   slerp2(double *, const double *, const double *,
-                        const double *, const double *, double, double);
+void  slerp1(float *, const float *, const float *, float);
+void  slerp2(float *, const float *, const float *,
+                      const float *, const float *, float, float);
 
 int extcmp(const char *, const char *);
